Add arrows to plot_process and label option to line plots
The new default behavior is to draw an arrow halfway along process
line plots pointing in the process direction.  The location of the
arrow as a percentage along the length of the line can be adjusted
with the arrow_pos between 0 and 1.  The arrow properties can be
adjusted with the arrowprops dictionary.

Optional labels can be added to all process and iso lines.  If a
string is passed to the label option, it will be added to the plot
as an annotation with the default behavior to be rotated to align
with the line at the label_pos as a percentage along on the length
of the line.

The angles will be incorrect if the axis limits or scale change
after they have been added.  Labels should be added after the axis
has been set to it's final settings.