update states.fix to accept property names as arguments
This update enables a shortened syntax for `states.fix()`.
If the independent properties have already been stored in
the `states` table, you can call the function with the
independent property column name as an argument and function
will look for the property at the state being fixed, which is
a more intuitive syntax. The previous style syntax will continue
to work as well.

>> from kilojoule.templates.default import *
>> water = realfluid.Properties('Water')
>> T[1] = Quantity(250,'degC')    # Temperature inlet 1
>> p[1] = Quantity(1,'MPa')
>> states.fix(1, T, p, water)